+++ src/app/data_harvester.rs
@@ -94,10 +94,10 @@ pub struct DataCollector {
 
 impl Default for DataCollector {
     fn default() -> Self {
-        trace!("Creating default data collector...");
+        // trace!("Creating default data collector...");
         DataCollector {
             data: Data::default(),
-            sys: System::new_all(),
+            sys: System::new_with_specifics(sysinfo::RefreshKind::new()),
             #[cfg(target_os = "linux")]
             pid_mapping: HashMap::new(),
             #[cfg(target_os = "linux")]
@@ -116,9 +116,10 @@ impl Default for DataCollector {
             battery_list: None,
             #[cfg(target_os = "linux")]
             page_file_size_kb: unsafe {
-                let page_file_size_kb = libc::sysconf(libc::_SC_PAGESIZE) as u64 / 1024;
-                trace!("Page file size in KB: {}", page_file_size_kb);
-                page_file_size_kb
+                // let page_file_size_kb = libc::sysconf(libc::_SC_PAGESIZE) as u64 / 1024;
+                // trace!("Page file size in KB: {}", page_file_size_kb);
+                // page_file_size_kb
+                libc::sysconf(libc::_SC_PAGESIZE) as u64 / 1024
             },
         }
     }
@@ -126,12 +127,13 @@ impl Default for DataCollector {
 
 impl DataCollector {
     pub fn init(&mut self) {
-        trace!("Initializing data collector.");
+        // trace!("Initializing data collector.");
+        self.sys.refresh_memory();
         self.mem_total_kb = self.sys.get_total_memory();
-        trace!("Total memory in KB: {}", self.mem_total_kb);
+        // trace!("Total memory in KB: {}", self.mem_total_kb);
 
         if self.widgets_to_harvest.use_battery {
-            trace!("First run battery vec creation.");
+            // trace!("First run battery vec creation.");
             if let Ok(battery_manager) = Manager::new() {
                 if let Ok(batteries) = battery_manager.batteries() {
                     let battery_list: Vec<Battery> = batteries.filter_map(Result::ok).collect();
@@ -143,15 +145,15 @@ impl DataCollector {
             }
         }
 
-        trace!("Running first run.");
+        // trace!("Running first run.");
         futures::executor::block_on(self.update_data());
-        trace!("First run done.  Sleeping for 250ms...");
+        // trace!("First run done.  Sleeping for 250ms...");
         std::thread::sleep(std::time::Duration::from_millis(250));
 
-        trace!("First run done.  Running first run cleanup now.");
+        // trace!("First run done.  Running first run cleanup now.");
         self.data.cleanup();
 
-        trace!("Enabled widgets to harvest: {:#?}", self.widgets_to_harvest);
+        // trace!("Enabled widgets to harvest: {:#?}", self.widgets_to_harvest);
     }
 
     pub fn set_collected_data(&mut self, used_widgets: UsedWidgets) {
@@ -208,13 +210,6 @@ impl DataCollector {
         // CPU
         if self.widgets_to_harvest.use_cpu {
             self.data.cpu = Some(cpu::get_cpu_data_list(&self.sys, self.show_average_cpu));
-            if log_enabled!(log::Level::Trace) {
-                if let Some(cpus) = &self.data.cpu {
-                    trace!("cpus: {:#?} results", cpus.len());
-                } else {
-                    trace!("Found no cpus.");
-                }
-            }
         }
 
         // Batteries
@@ -222,14 +217,6 @@ impl DataCollector {
             if let Some(battery_list) = &mut self.battery_list {
                 self.data.list_of_batteries =
                     Some(batteries::refresh_batteries(&battery_manager, battery_list));
-            }
-
-            if log_enabled!(log::Level::Trace) {
-                if let Some(batteries) = &self.data.list_of_batteries {
-                    trace!("batteries: {:#?} results", batteries.len());
-                } else {
-                    trace!("Found no batteries.");
-                }
             }
         }
 
@@ -259,14 +246,6 @@ impl DataCollector {
                 }
             } {
                 self.data.list_of_processes = Some(process_list);
-            }
-
-            if log_enabled!(log::Level::Trace) {
-                if let Some(processes) = &self.data.list_of_processes {
-                    trace!("processes: {:#?} results", processes.len());
-                } else {
-                    trace!("Found no processes.");
-                }
             }
         }
 
@@ -362,63 +341,26 @@ impl DataCollector {
             self.total_rx = net_data.total_rx;
             self.total_tx = net_data.total_tx;
             self.data.network = Some(net_data);
-            if log_enabled!(log::Level::Trace) {
-                trace!("Total rx: {:#?}", self.total_rx);
-                trace!("Total tx: {:#?}", self.total_tx);
-                if let Some(network) = &self.data.network {
-                    trace!("network rx: {:#?}", network.rx);
-                    trace!("network tx: {:#?}", network.tx);
-                } else {
-                    trace!("Could not find any networks.");
-                }
-            }
         }
 
         if let Ok(memory) = mem_res.0 {
             self.data.memory = memory;
-            if log_enabled!(log::Level::Trace) {
-                trace!("mem: {:?} results", self.data.memory);
-            }
         }
 
         if let Ok(swap) = mem_res.1 {
             self.data.swap = swap;
-            if log_enabled!(log::Level::Trace) {
-                trace!("swap: {:?} results", self.data.swap);
-            }
         }
 
         if let Ok(disks) = disk_res {
             self.data.disks = disks;
-            if log_enabled!(log::Level::Trace) {
-                if let Some(disks) = &self.data.disks {
-                    trace!("disks: {:#?} results", disks.len());
-                } else {
-                    trace!("Could not find any disks.");
-                }
-            }
         }
 
         if let Ok(io) = io_res {
             self.data.io = io;
-            if log_enabled!(log::Level::Trace) {
-                if let Some(io) = &self.data.io {
-                    trace!("io: {:#?} results", io.len());
-                } else {
-                    trace!("Could not find any io results.");
-                }
-            }
         }
 
         if let Ok(temp) = temp_res {
             self.data.temperature_sensors = temp;
-            if log_enabled!(log::Level::Trace) {
-                if let Some(sensors) = &self.data.temperature_sensors {
-                    trace!("temp: {:#?} results", sensors.len());
-                } else {
-                    trace!("Could not find any temp sensors.");
-                }
-            }
         }
 
         // Update time