+++ src/bin/main.rs
@@ -27,27 +27,24 @@ use tui::{backend::CrosstermBackend, Terminal};
 
 fn main() -> Result<()> {
     let matches = clap::get_matches();
-    let is_debug = matches.is_present("debug");
-    if is_debug {
-        let mut tmp_dir = std::env::temp_dir();
-        tmp_dir.push("bottom_debug.log");
-        utils::logging::init_logger(log::LevelFilter::Trace, tmp_dir.as_os_str())?;
-    } else {
-        #[cfg(debug_assertions)]
-        {
-            utils::logging::init_logger(
-                log::LevelFilter::Debug,
-                std::ffi::OsStr::new("debug.log"),
-            )?;
-        }
+    // let is_debug = matches.is_present("debug");
+    // if is_debug {
+    //     let mut tmp_dir = std::env::temp_dir();
+    //     tmp_dir.push("bottom_debug.log");
+    //     utils::logging::init_logger(log::LevelFilter::Trace, tmp_dir.as_os_str())?;
+    // } else {
+    #[cfg(debug_assertions)]
+    {
+        utils::logging::init_logger(log::LevelFilter::Debug, std::ffi::OsStr::new("debug.log"))?;
     }
+    // }
 
     let config_path = read_config(matches.value_of("config_location"))
         .context("Unable to access the given config file location.")?;
-    trace!("Config path: {:?}", config_path);
+    // trace!("Config path: {:?}", config_path);
     let mut config: Config = create_or_get_config(&config_path)
         .context("Unable to properly parse or create the config file.")?;
-    trace!("Current config: {:#?}", config);
+    // trace!("Current config: {:#?}", config);
 
     // Get widget layout separately
     let (widget_layout, default_widget_id, default_widget_type_option) =
@@ -87,29 +84,31 @@ fn main() -> Result<()> {
         let lock = thread_termination_lock.clone();
         let cvar = thread_termination_cvar.clone();
         let cleaning_sender = sender.clone();
-        trace!("Initializing cleaning thread...");
+        const OFFSET_WAIT_TIME: u64 = constants::STALE_MAX_MILLISECONDS + 60000;
+        // trace!("Initializing cleaning thread...");
         thread::spawn(move || {
             loop {
+                // debug!("Starting cleaning loop...");
                 let result = cvar.wait_timeout(
                     lock.lock().unwrap(),
-                    Duration::from_millis(constants::STALE_MAX_MILLISECONDS + 5000),
+                    Duration::from_millis(OFFSET_WAIT_TIME),
                 );
+                // debug!("Result mutex guard over...");
                 if let Ok(result) = result {
                     if *(result.0) {
-                        trace!("Received termination lock in cleaning thread from cvar!");
+                        // debug!("Received termination lock in cleaning thread from cvar!");
                         break;
                     }
-                } else {
-                    trace!("Sending cleaning signal...");
-                    if cleaning_sender.send(BottomEvent::Clean).is_err() {
-                        trace!("Failed to send cleaning signal.  Halting cleaning thread loop.");
-                        break;
-                    }
-                    trace!("Cleaning signal sent without errors.");
                 }
+                // debug!("Sending cleaning signal...");
+                if cleaning_sender.send(BottomEvent::Clean).is_err() {
+                    // debug!("Failed to send cleaning sender...");
+                    break;
+                }
+                // trace!("Cleaning signal sent without errors.");
             }
 
-            trace!("Cleaning thread loop has closed.");
+            // trace!("Cleaning thread loop has closed.");
         })
     };
 
@@ -146,13 +145,6 @@ fn main() -> Result<()> {
 
     while !is_terminated.load(Ordering::SeqCst) {
         if let Ok(recv) = receiver.recv_timeout(Duration::from_millis(TICK_RATE_IN_MILLISECONDS)) {
-            if log_enabled!(log::Level::Trace) {
-                if let BottomEvent::Update(_) = recv {
-                    trace!("Main/drawing thread received Update event.");
-                } else {
-                    trace!("Main/drawing thread received event: {:?}", recv);
-                }
-            }
             match recv {
                 BottomEvent::KeyInput(event) => {
                     if handle_key_event_or_break(event, &mut app, &collection_thread_ctrl_sender) {
@@ -247,18 +239,18 @@ fn main() -> Result<()> {
         }
 
         // TODO: [OPT] Should not draw if no change (ie: scroll max)
-        try_drawing(&mut terminal, &mut app, &mut painter, is_debug)?;
+        try_drawing(&mut terminal, &mut app, &mut painter)?;
     }
 
     // I think doing it in this order is safe...
-    trace!("Send termination thread locks.");
+    // trace!("Send termination thread locks.");
     *thread_termination_lock.lock().unwrap() = true;
-    trace!("Notifying all cvars.");
+    // trace!("Notifying all cvars.");
     thread_termination_cvar.notify_all();
 
-    trace!("Main/drawing thread is cleaning up.");
-    cleanup_terminal(&mut terminal, is_debug)?;
+    // trace!("Main/drawing thread is cleaning up.");
+    cleanup_terminal(&mut terminal)?;
 
-    trace!("Fini.");
+    // trace!("Fini.");
     Ok(())
 }